requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
numpy==1.16.3
python-Levenshtein==0.12.0
six==1.12.0
ujson==1.35
Werkzeug==0.15.4
WTForms==2.2.1
xpinyin==0.5.6
xgboost==0.90
pandas==0.23.0
pyahocorasick==1.1.6
pyltp==0.2.1
zhon==1.1.5
cn2an==0.4.1
regex==2019.6.8
arrow==0.14.2
lxml==4.3.4
